New separate .js files; main.js connected as module js file;
