+++ js/data.js
@@ -1,0 +1,29 @@
+const POSSIBLE_NAMES = [
+  'Дмитрий',
+  'Микола',
+  'Клементина',
+  'Деметриус',
+  'Виктория',
+  'user_unknown',
+];
+
+const POSSIBLE_DESCRIPTIONS = [
+  'Хорошая фотография?',
+  'Я решил заняться фотографией!',
+  'Мяу!',
+  'Как вам фото?'
+];
+
+const POSSIBLE_MESSAGE = [
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.'
+];
+
+const usedCommentsID = [];
+
+export {POSSIBLE_NAMES, POSSIBLE_DESCRIPTIONS, POSSIBLE_MESSAGE, usedCommentsID};
+
